refactor: calculate_time.cをtime.cに移動
