+++ philo/Makefile
@@ -20,7 +20,6 @@ SRCS = \
 	src/bool_handlers.c \
 	src/fork.c \
 	src/change_data.c \
-	src/calculate_time.c \
 	src/free.c \
 	src/monitor.c \
 	src/is_number.c \
